+++ api/src/main/java/org/allaymc/api/block/component/BlockBlockEntityHolderComponent.java
@@ -17,34 +17,21 @@ public interface BlockBlockEntityHolderComponent<T extends BlockEntity> extends 
     BlockEntityType<?> getBlockEntityType();
 
     /**
-     * Get the block entity in a specific location.
-     *
-     * @param x         block entity's x coordinate
-     * @param y         block entity's y coordinate
-     * @param z         block entity's z coordinate
-     * @param dimension the dimension which the block entity is in
-     * @return the block entity, or null if block entity is not found
-     */
-    default T getBlockEntity(int x, int y, int z, Dimension dimension) {
-        var blockEntity = dimension.getBlockEntity(x, y, z);
-        if (blockEntity == null) return null;
-        if (blockEntity.getBlockEntityType() != this.getBlockEntityType()) {
-            throw new IllegalStateException(
-                    "Mismatched block entity type at pos %d, %d, %d, %s! Expected: %s, actual: %s"
-                            .formatted(
-                                    x, y, z, dimension.toString(),
-                                    this.getBlockEntityType().getName(), blockEntity.getBlockEntityType().getName()
-                            )
-            );
-        }
-        return (T) blockEntity;
-    }
-
-    /**
      * @see #getBlockEntity(int, int, int, Dimension)
      */
     default T getBlockEntity(Position3ic pos) {
         return getBlockEntity(pos.x(), pos.y(), pos.z(), pos.dimension());
     }
 
+    /**
+     * Get the block entity in a specific location. If the block entity is not present, it will be created.
+     *
+     * @param x         block entity's x coordinate
+     * @param y         block entity's y coordinate
+     * @param z         block entity's z coordinate
+     * @param dimension the dimension which the block entity is in
+     * @return the block entity in the specified location
+     * @throws IllegalStateException if the block entity is not the correct type
+     */
+    T getBlockEntity(int x, int y, int z, Dimension dimension);
 }