+++ api/src/main/java/org/allaymc/api/block/interfaces/BlockNoteblockBehavior.java
@@ -1,6 +1,8 @@
 package org.allaymc.api.block.interfaces;
 
 import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.component.BlockBlockEntityHolderComponent;
+import org.allaymc.api.blockentity.interfaces.BlockEntityNoteblock;
 
-public interface BlockNoteblockBehavior extends BlockBehavior {
+public interface BlockNoteblockBehavior extends BlockBehavior, BlockBlockEntityHolderComponent<BlockEntityNoteblock> {
 }