+++ api/src/main/java/org/allaymc/api/blockentity/interfaces/BlockEntityNoteblock.java
@@ -1,0 +1,10 @@
+package org.allaymc.api.blockentity.interfaces;
+
+import org.allaymc.api.blockentity.BlockEntity;
+import org.allaymc.api.blockentity.component.BlockEntityNoteblockBaseComponent;
+
+/**
+ * @author daoge_cmd
+ */
+public interface BlockEntityNoteblock extends BlockEntity, BlockEntityNoteblockBaseComponent {
+}